+++ src/components/orchestrator/Orchestrator.test.tsx
@@ -1,17 +1,29 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { useSearch } from '@tanstack/react-router'
 import { act, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { expect } from 'vitest'
 
 import { MODE_TYPE } from '@/constants/mode'
+import { PAGE_TYPE } from '@/constants/page'
 import { TELEMETRY_EVENT_TYPE } from '@/constants/telemetry'
 import { TelemetryContext } from '@/contexts/TelemetryContext'
+import type { Interrogation } from '@/models/interrogation'
 import { renderWithRouter } from '@/utils/tests'
 
 import { Orchestrator } from './Orchestrator'
 
+vi.mock('@tanstack/react-router', async (importOriginal) => {
+  const actual =
+    (await importOriginal()) as typeof import('@tanstack/react-router')
+  return {
+    ...actual,
+    useSearch: vi.fn(),
+  }
+})
+
 describe('Orchestrator', () => {
-  const interrogation = {
+  const defaultInterrogation = {
     stateData: undefined,
     data: undefined,
     personalization: undefined,
@@ -50,30 +62,33 @@ describe('Orchestrator', () => {
     maxPage: '2',
   }
   const queryClient = new QueryClient()
+
+  const mockGetDepositProof = vi.fn()
+
   const OrchestratorTestWrapper = ({
     mode,
     isDownloadEnabled = false,
+    initialInterrogation = defaultInterrogation,
   }: {
     mode: MODE_TYPE.COLLECT | MODE_TYPE.REVIEW | MODE_TYPE.VISUALIZE
     isDownloadEnabled?: boolean
+    initialInterrogation?: Interrogation
   }) => (
     <QueryClientProvider client={queryClient}>
       <Orchestrator
         metadata={metadata}
         mode={mode}
         isDownloadEnabled={isDownloadEnabled}
-        initialInterrogation={interrogation}
+        initialInterrogation={initialInterrogation}
         // @ts-expect-error: we should have a better lunatic mock
         source={source}
         getReferentiel={() => {
           return new Promise(() => [])
         }}
         updateDataAndStateData={() => {
-          return new Promise<void>(() => {})
+          return new Promise<void>(() => { })
         }}
-        getDepositProof={() => {
-          return new Promise<void>(() => {})
-        }}
+        getDepositProof={mockGetDepositProof}
       />
     </QueryClientProvider>
   )
@@ -85,7 +100,7 @@ describe('Orchestrator', () => {
       <TelemetryContext.Provider
         value={{
           isTelemetryEnabled: true,
-          pushEvent: () => {},
+          pushEvent: () => { },
           setDefaultValues,
         }}
       >
@@ -109,7 +124,7 @@ describe('Orchestrator', () => {
         value={{
           isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -134,7 +149,7 @@ describe('Orchestrator', () => {
         value={{
           isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.VISUALIZE} />
@@ -154,7 +169,7 @@ describe('Orchestrator', () => {
         value={{
           isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.REVIEW} />
@@ -174,7 +189,7 @@ describe('Orchestrator', () => {
         value={{
           isTelemetryEnabled: false,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -194,7 +209,7 @@ describe('Orchestrator', () => {
         value={{
           isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -220,7 +235,7 @@ describe('Orchestrator', () => {
         value={{
           isTelemetryEnabled: true,
           pushEvent,
-          setDefaultValues: () => {},
+          setDefaultValues: () => { },
         }}
       >
         <OrchestratorTestWrapper mode={MODE_TYPE.COLLECT} />
@@ -310,5 +325,34 @@ describe('Orchestrator', () => {
     )
     act(() => getByText('Start').click())
     expect(queryByText('Download data')).toBeNull()
+  })
+
+  it('calls getDepositProof with right params', async () => {
+    // Set initial interrogation as validated on end page to directly enable download receipt
+    const initialInterrogation: Interrogation = {
+      ...defaultInterrogation,
+      stateData: { state: 'VALIDATED', currentPage: PAGE_TYPE.END, date: 1 },
+    }
+
+    // Provide encoded param via URL
+    vi.mocked(useSearch).mockReturnValue({
+      surveyUnitCompositeName: 'RnJhbsOnYWlz',
+    })
+
+    const { getByText } = renderWithRouter(
+      <OrchestratorTestWrapper
+        mode={MODE_TYPE.COLLECT}
+        initialInterrogation={initialInterrogation}
+      />,
+    )
+
+    vi.mocked(mockGetDepositProof).mockResolvedValue(undefined)
+
+    // download deposit proof
+    act(() => getByText('Download the receipt').click())
+
+    expect(mockGetDepositProof).toHaveBeenCalledWith({
+      surveyUnitCompositeName: 'RnJhbsOnYWlz',
+    })
   })
 })