+++ src/pages/collect/CollectPage.tsx
@@ -12,6 +12,7 @@ import {
 import { showToast } from '@/components/Toast'
 import { Orchestrator } from '@/components/orchestrator/Orchestrator'
 import { MODE_TYPE } from '@/constants/mode'
+import type { GenerateDepositProofParams } from '@/models/api'
 import type { LunaticGetReferentiel, Nomenclature } from '@/models/lunaticType'
 import type { StateData } from '@/models/stateData'
 
@@ -76,9 +77,11 @@ export const CollectPage = memo(function CollectPage() {
         }
       })
 
-  const getDepositProof = () =>
+  const getDepositProof = (params?: GenerateDepositProofParams) =>
     queryClient
-      .ensureQueryData(getGenerateDepositProofQueryOptions(interrogationId))
+      .ensureQueryData(
+        getGenerateDepositProofQueryOptions(interrogationId, params),
+      )
       .then((response) => {
         const fileName =
           (response.headers['content-disposition']?.match(